Very mis-leading!
This is very misleading.... Setting your servers flags to Aikar's is not a magical fix for lag. The timings website should not be telling users that this is a fix for lag.

Whilst Aikar's flags can be a fix for some lag, it is mis-leading in the fact that this is the first thing that people see when visiting the timings website. Instead, maybe have a look at other issues in your timings report?

Please consider changing this text to something not so mis-leading.